support MOE

+++ tests/huggingface_test_runner.py
@@ -77,7 +77,9 @@ def dequantize_weights(model_dir):
         if filename.endswith(".safetensors") and not filename.endswith(".org.safetensors"):
             filepath = os.path.join(model_dir, filename)
             org_filepath = filepath.replace(".safetensors", ".org.safetensors")
-
+            if filename.endswith(".safetensors") and os.path.exists(os.path.join(model_dir, org_filepath)):
+                continue
+        
             if not os.path.exists(org_filepath):
                 os.rename(filepath, org_filepath)
 
@@ -169,13 +171,13 @@ class HuggingfaceTestRunner(TestRunner):
             count = 0
             if (self.cfg['huggingface_options']['output_logits']):
                 if not test_utils.in_ci():
-                    logits = result.logits.detach().numpy()[0]
+                    logits = result.logits.detach().to(torch.float32).numpy()[0]
                     dump_data_to_file(self.case_dir, f'cpu_result_{count}', logits)
                     outputs.append(logits)
                     count += 1
             else:
                 if not test_utils.in_ci():
-                    hidden_states = recursive_stack(result.hidden_states).detach().numpy()[-1][0]
+                    hidden_states = recursive_stack(result.hidden_states).detach().to(torch.float32).numpy()[-1][0]
                     dump_data_to_file(self.case_dir, f'cpu_result_{count}', hidden_states)
                     outputs.append(hidden_states)
                     count += 1
@@ -236,12 +238,12 @@ class HuggingfaceTestRunner(TestRunner):
 
         self.cfg['huggingface_options']['config'] = self.kv_cache_config
 
-        if hasattr(config, "quantization_config"):
-            dequantize_weights(model_path)
-            delattr(config, "quantization_config")
+        # if hasattr(config, "quantization_config"):
+        #     dequantize_weights(model_path)
+        #     delattr(config, "quantization_config")
         self.model = AutoModelForCausalLM.from_pretrained(
-            model_path, config=config, torch_dtype="auto", device_map="cpu", trust_remote_code=True).to(torch.float32).eval()
-        restore_weights(model_path)
+            model_path, config=config, torch_dtype="auto", device_map="auto", trust_remote_code=True).eval()
+        # restore_weights(model_path)
         self.tokenizer = AutoTokenizer.from_pretrained(model_path, trust_remote_code=True)
         self.generation_config = self.model.generation_config
         # self.generation_config.return_dict_in_generate = True # if False, generate only output tokens
